# document/notification/api-docs.json
{"openapi":"3.1.0","info":{"title":"EcomHub API","version":"v1"},"servers":[{"url":"http://localhost:8081","description":"Development"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Thông báo","description":"API quản lý thông báo - xem, đếm và quản lý trạng thái thông báo"},{"name":"Mẫu thông báo","description":"API quản lý mẫu thông báo - tạo, xem, cập nhật và tìm kiếm mẫu thông báo"}],"paths":{"/api/v1/templates":{"get":{"tags":["Mẫu thông báo"],"summary":"Lấy tất cả mẫu thông báo","description":"Lấy danh sách tất cả mẫu thông báo trong hệ thống","operationId":"getAll","responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}}}},"post":{"tags":["Mẫu thông báo"],"summary":"Tạo mẫu thông báo mới","description":"Tạo một mẫu thông báo mới với thông tin chi tiết","operationId":"add","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddTemplateRequest"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Dữ liệu yêu cầu không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Dữ liệu yêu cầu không hợp lệ\",\n  \"data\": null\n}\n"}}},"409":{"description":"Mẫu thông báo với tên này đã tồn tại","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 409,\n  \"errorCode\": \"TEMPLATE_ALREADY_EXISTS\",\n  \"message\": \"Mẫu thông báo với tên này đã tồn tại\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/templates/{name}/update-subject":{"patch":{"tags":["Mẫu thông báo"],"summary":"Cập nhật tiêu đề mẫu thông báo","description":"Cập nhật tiêu đề của một mẫu thông báo hiện có","operationId":"updateSubject","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTemplateSubjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Tên mẫu thông báo không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Tên mẫu thông báo không hợp lệ\",\n  \"data\": null\n}\n"}}},"404":{"description":"Không tìm thấy mẫu thông báo với tên đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy mẫu thông báo với tên đã cho\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/templates/{name}/update-name":{"patch":{"tags":["Mẫu thông báo"],"summary":"Cập nhật tên mẫu thông báo","description":"Cập nhật tên của một mẫu thông báo hiện có","operationId":"updateName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTemplateNameRequest"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Tên mẫu thông báo không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Tên mẫu thông báo không hợp lệ\",\n  \"data\": null\n}\n"}}},"404":{"description":"Không tìm thấy mẫu thông báo với tên đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy mẫu thông báo với tên đã cho\",\n  \"data\": null\n}\n"}}},"409":{"description":"Tên mẫu thông báo mới đã tồn tại","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 409,\n  \"errorCode\": \"RESOURCE_ALREADY_EXISTS\",\n  \"message\": \"Tên mẫu thông báo mới đã tồn tại\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/templates/{name}/update-description":{"patch":{"tags":["Mẫu thông báo"],"summary":"Cập nhật mô tả mẫu thông báo","description":"Cập nhật mô tả của một mẫu thông báo hiện có","operationId":"updateDescription","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTemplateDescriptionRequest"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Tên mẫu thông báo không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Tên mẫu thông báo không hợp lệ\",\n  \"data\": null\n}\n"}}},"404":{"description":"Không tìm thấy mẫu thông báo với tên đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy mẫu thông báo với tên đã cho\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/templates/{name}/update-body":{"patch":{"tags":["Mẫu thông báo"],"summary":"Cập nhật nội dung mẫu thông báo","description":"Cập nhật nội dung chính của một mẫu thông báo hiện có","operationId":"updateBody","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTemplateBodyRequest"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Tên mẫu thông báo không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Tên mẫu thông báo không hợp lệ\",\n  \"data\": null\n}\n"}}},"404":{"description":"Không tìm thấy mẫu thông báo với tên đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy mẫu thông báo với tên đã cho\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/notifications/{notificationId}/read":{"patch":{"tags":["Thông báo"],"summary":"Đánh dấu thông báo đã đọc","description":"Đánh dấu một thông báo cụ thể là đã đọc","operationId":"markAsRead","parameters":[{"name":"notificationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"404":{"description":"Không tìm thấy thông báo với ID đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy thông báo với ID đã cho\",\n  \"data\": null\n}\n"}}},"500":{"description":"Cập nhật thông tin không thành công","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 500,\n  \"errorCode\": \"UPDATE_FAILED\",\n  \"message\": \"Cập nhật thông tin không thành công\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/notifications/read-all":{"patch":{"tags":["Thông báo"],"summary":"Đánh dấu tất cả thông báo đã đọc","description":"Đánh dấu tất cả thông báo của người dùng hiện tại là đã đọc","operationId":"markAllAsRead","responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"404":{"description":"Không tìm thấy người dùng với ID đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"USER_NOT_FOUND\",\n  \"message\": \"Không tìm thấy người dùng với ID đã cho\",\n  \"data\": null\n}\n"}}},"500":{"description":"Cập nhật thông tin không thành công","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 500,\n  \"errorCode\": \"UPDATE_FAILED\",\n  \"message\": \"Cập nhật thông tin không thành công\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/templates/search":{"get":{"tags":["Mẫu thông báo"],"summary":"Tìm kiếm mẫu thông báo theo tên","description":"Tìm kiếm mẫu thông báo dựa trên tên mẫu","operationId":"searchByName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}}}}},"/api/v1/templates/channel/{channelName}":{"get":{"tags":["Mẫu thông báo"],"summary":"Lấy mẫu thông báo theo kênh","description":"Lấy danh sách tất cả mẫu thông báo của một kênh cụ thể","operationId":"getByChannel","parameters":[{"name":"channelName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"400":{"description":"Tên kênh không hợp lệ","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 400,\n  \"errorCode\": \"REQUIRED_FIELD_MISSING\",\n  \"message\": \"Tên kênh không hợp lệ\",\n  \"data\": null\n}\n"}}},"404":{"description":"Không tìm thấy kênh với tên đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy kênh với tên đã cho\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/notifications/recipient/{recipientId}":{"get":{"tags":["Thông báo"],"summary":"Lấy danh sách thông báo theo người nhận","description":"Lấy danh sách thông báo của một người nhận cụ thể với khả năng lọc theo thời gian tạo và giới hạn số lượng","operationId":"getByRecipientId","parameters":[{"name":"recipientId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"createAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{}}}}}},"404":{"description":"Không tìm thấy người nhận với ID đã cho","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"example":"{\n  \"statusCode\": 404,\n  \"errorCode\": \"RESOURCE_NOT_FOUND\",\n  \"message\": \"Không tìm thấy người nhận với ID đã cho\",\n  \"data\": null\n}\n"}}}}}},"/api/v1/notifications/recipient/count-unread":{"get":{"tags":["Thông báo"],"summary":"Đếm số thông báo chưa đọc","description":"Đếm số lượng thông báo chưa đọc của một người nhận cụ thể","operationId":"countUnread","responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"properties":{"statusCode":{"type":"integer","format":"int32"},"errorCode":{"type":"string"},"message":{"type":"string"},"data":{"type":"integer","format":"int32"}}}}}}}}}},"components":{"schemas":{"AddTemplateRequest":{"type":"object","properties":{"name":{"type":"string"},"channel":{"type":"string"},"subjectTemplate":{"type":"string"},"bodyTemplate":{"type":"string"},"description":{"type":"string"}}},"UpdateTemplateSubjectRequest":{"type":"object","properties":{"newSubject":{"type":"string"}}},"UpdateTemplateNameRequest":{"type":"object","properties":{"newName":{"type":"string"}}},"UpdateTemplateDescriptionRequest":{"type":"object","properties":{"newDescription":{"type":"string"}}},"UpdateTemplateBodyRequest":{"type":"object","properties":{"newBody":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}